backend/requirements.txt
fastapi==0.135.1
uvicorn==0.41.0
openai==2.24.0
python-dotenv==1.2.2
requests==2.32.5
pydantic==2.12.5
httpx==0.28.1
